requirements.txt
playwright
beautifulsoup4
pandas
anthropic
python-dotenv
sqlalchemy
schedule
openpyxl
